Replace double floats with single floats

+++ src/anneal-ocl.h
@@ -6,11 +6,11 @@ struct an_gpu_context;
 struct an_image;
 
 /* Fourier transform */
-int an_rfft (const cl_uchar        *array,
-             cl_double             *real,
-             cl_double             *imag,
-             const cl_uint         *dimensions,
-             unsigned int           ndims);
+int an_rfft (const cl_uchar *array,
+             cl_float       *real,
+             cl_float       *imag,
+             const cl_uint  *dimensions,
+             unsigned int    ndims);
 
 /* GPU context */
 struct an_gpu_context*
@@ -22,14 +22,14 @@ an_destroy_gpu_context (struct an_gpu_context *ctx);
 /* Image object */
 struct an_image*
 an_create_image (struct an_gpu_context *ctx,
-                 const cl_double       *real,
-                 const cl_double       *imag,
+                 const cl_float        *real,
+                 const cl_float        *imag,
                  const cl_uint         *dimensions,
                  unsigned int           ndims);
 
 struct an_image*
 an_create_corrfn (struct an_gpu_context *ctx,
-                  const cl_double       *corrfn,
+                  const cl_float        *corrfn,
                   const cl_uint         *dimensions,
                   unsigned int           ndims);
 
@@ -46,4 +46,4 @@ an_image_update_fft (struct an_image *image,
 int
 an_distance (struct an_image *target,
              struct an_image *recon,
-             cl_double       *distance);
+             cl_float        *distance);
